Move insert and slice methods into StringUtils class.
